+++ R/defaults.R
@@ -23,6 +23,11 @@ default_flextable_settings <- list(
   na_str  = "",
   nan_str  = "",
   fmt_date = "%Y-%m-%d", fmt_datetime = "%Y-%m-%d %H:%M:%S",
+
+  htmlscroll = FALSE, shadow = FALSE,
+  split = TRUE, keep_with_next = FALSE,
+  tabcolsep = 0, arraystretch = 1.5, float = "none",
+
   fonts_ignore = FALSE,
   extra_css = "",
   theme_fun = "theme_booktabs",
@@ -76,6 +81,27 @@ flextable_global$defaults <- default_flextable_settings
 #' xelatex or lualatex. If pdflatex is used, fonts will be ignored because they are
 #' not supported by pdflatex, whereas with the xelatex and lualatex engines they are.
 #' @param extra_css css instructions to be integrated with the table.
+#' @param htmlscroll `TRUE` or `FALSE`, enable horizontal scrolling.
+#' @param shadow `TRUE` or `FALSE`, use shadow dom (for HTML only), this option is existing
+#' to disable shadow dom (set to `FALSE`) for pagedown and Quarto that can
+#' not support it for now.
+#' @param split Word option 'Allow row to break across pages' can be
+#' activated when TRUE.
+#' @param keep_with_next Word option 'keep rows together' is
+#' activated when TRUE. It avoids page break within tables. This
+#' is handy for small tables, i.e. less than a page height.
+#' @param tabcolsep space between the text and the left/right border of its containing
+#' cell.
+#' @param arraystretch height of each row relative to its default
+#' height, the default value is 1.5.
+#' @param float type of floating placement in the PDF document, one of:
+#' * 'none' (the default value), table is placed after the preceding
+#' paragraph.
+#' * 'float', table can float to a place in the text where it fits best
+#' * 'wrap-r', wrap text around the table positioned to the right side of the text
+#' * 'wrap-l', wrap text around the table positioned to the left side of the text
+#' * 'wrap-i', wrap text around the table positioned inside edge-near the binding
+#' * 'wrap-o', wrap text around the table positioned outside edge-far from the binding
 #' @param theme_fun a single character value (the name of the theme function
 #' to be applied) or a theme function (input is a flextable, output is a flextable).
 #' @param post_process_pdf,post_process_docx,post_process_html,post_process_pptx Post-processing functions
@@ -113,8 +139,13 @@ set_flextable_defaults <- function(
   cs.family = NULL, eastasia.family = NULL, hansi.family = NULL,
   decimal.mark = NULL, big.mark = NULL, digits = NULL,
   na_str = NULL, nan_str = NULL,
-  fmt_date = NULL, fmt_datetime = NULL, extra_css = NULL,
-  fonts_ignore = NULL, theme_fun = NULL,
+  fmt_date = NULL, fmt_datetime = NULL,
+  extra_css = NULL,
+  htmlscroll = NULL, shadow = NULL,
+  split = NULL, keep_with_next = NULL,
+  tabcolsep = NULL, arraystretch = NULL, float = NULL,
+  fonts_ignore = NULL,
+  theme_fun = NULL,
   post_process_pdf = NULL,
   post_process_docx = NULL,
   post_process_html = NULL,
@@ -198,6 +229,28 @@ set_flextable_defaults <- function(
   if( !is.null(fonts_ignore) ){
     x$fonts_ignore <- fonts_ignore
   }
+  if( !is.null(htmlscroll) ){
+    x$htmlscroll <- htmlscroll
+  }
+  if( !is.null(shadow) ){
+    x$shadow <- shadow
+  }
+  if( !is.null(split) ){
+    x$split <- split
+  }
+  if( !is.null(keep_with_next) ){
+    x$keep_with_next <- keep_with_next
+  }
+  if( !is.null(tabcolsep) ){
+    x$tabcolsep <- tabcolsep
+  }
+  if( !is.null(arraystretch) ){
+    x$arraystretch <- arraystretch
+  }
+  if( !is.null(float) ){
+    x$float <- float
+  }
+
   if( !is.null(extra_css) ){
     x$extra_css <- extra_css
   }